--- a/src/GraphQL/Execution/ReadonlyResolveFieldContextPolicy.cs
+++ b/src/GraphQL/Execution/ReadonlyResolveFieldContextPolicy.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.ObjectPool;
+
+namespace GraphQL.Execution
+{
+    internal sealed class ReadonlyResolveFieldContextPolicy : IPooledObjectPolicy<ReadonlyResolveFieldContext>
+    {
+        public ReadonlyResolveFieldContext Create() => new ReadonlyResolveFieldContext();
+
+        public bool Return(ReadonlyResolveFieldContext obj)
+        {
+            obj.Clear();
+            return true;
+        }
+    }
+}