Slight optimization over `ReadonlyResolveFieldContext`. No breaking changes, no public api changes, only internals. GraphQL package gets new dependency from `Microsoft.Extensions.ObjectPool` (netstandard2.0). I find it acceptable for internal things. One could do the cache himself, but why if there is a ready-made solution?

Before
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.17134.1304 (1803/April2018Update/Redstone4)
Intel Core i5-6200U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
Frequency=2343753 Hz, Resolution=426.6661 ns, Timer=TSC
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X86 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X86 RyuJIT


```
|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 2,607.83 us | 26.774 us | 23.735 us | 195.3125 | 54.6875 |     - | 486.71 KB |
|          Hero |    57.52 us |  0.308 us |  0.257 us |   8.8501 |       - |     - |  13.61 KB |

After
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.17134.1304 (1803/April2018Update/Redstone4)
Intel Core i5-6200U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
Frequency=2343753 Hz, Resolution=426.6661 ns, Timer=TSC
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X86 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X86 RyuJIT


```
|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 2,643.91 us | 14.938 us | 13.973 us | 171.8750 | 50.7813 |     - | 454.89 KB |
|          Hero |    58.63 us |  0.341 us |  0.319 us |   8.7891 |       - |     - |  13.55 KB |

~7% memory decrease on introspection test. There is some slowdown but it is not so significant.

@Shane32 Perhaps you will find the use of the pool in other places.