@Shane32 2020-03-03
> No breaking changes, no public api changes, only internals.

The context is exposed outside of the internals of the project.  The execution strategy can only assume it has been discarded, so it would be false to say that it was not a breaking change.  For instance, if someone utilized an UnhandledExceptionDelegate to store references to the context of failed resolvers, saving values from them in the database via a DocumentListener after execution was complete, the contexts would contain incorrect values.

Interestingly enough, some of my own code in production would break (in two different spots) if the ObjectPool was implemented in this manner.

> I find it acceptable for internal things. 

I would agree, and suggest that we only utilize the ObjectPool where the object is not exposed, directly or indirectly, to the user.  This is clearly not one of those scenarios.

> GraphQL package gets new dependency

Personally, I do not have a problem with a dependency, so long as it does not cause incompatibility with prior versions of .Net.  Note that this package is not part of .Net Standard 2.0, but rather is a part of Asp.Net Core.

According to Microsoft: 
> Microsoft.Extensions.ObjectPool is part of the ASP.NET Core infrastructure
> https://docs.microsoft.com/en-us/aspnet/core/performance/objectpool?view=aspnetcore-3.1

However, it only has a single dependency, and that is .Net Standard 2.0.  We should at least verify that applications that rely on .Net Core 2.2 (like mine) should be able to safely upgrade that one nuget package without breaking compatibility with the rest of their application.  I could probably test this if necessary.

Personally, I'd import the ObjectPool version 2.0 nuget package and then allow compatibility with >= 2.0 so it works without changing a user's dependency tree, at least as far back as Asp.Net Core 2.0.  If a person wanted the latest 3.1.1 version, they can always import that (if they do not already have it imported automatically when they import Asp.Net Core 3.1).

I'm also confused why you don't seem to mind this dependency, whereas you've called other dependencies "leaky", even when they are only abstractions and contain no code.  Again, I do not mind dependencies so long as they work properly with older versions of Asp.Net Core.  (It's a very big problem when one nuget package requires the very latest dependencies of everything for no good reason.)

> ~7% memory decrease on introspection test. There is some slowdown but it is not so significant.

For the sake of discussion, let's compare this to #1535 .  With the latter, there is no performance drop, and has the same GC drop.  Memory usage drops 4% instead of 7%.  There is no negative side effects.  It would even be tough to call #1535 "breaking" if it wasn't for the fact that ReadonlyResolveFieldContext is public, added in the last preview build.  By simply not removing the class, it wouldn't even be breaking.

I could not justify adding a dependency and adding known negative side effects for a minor memory gain (and performance loss), when a second alternative has a similar effect without the negative consequences.  Even if there was no alternative, I would not pull it in because of its unique requirements on user code (namely, that the context is only valid during the execution of the resolver, middleware, error handlers, etc).

Also: my data loader PR #1511 will not work with the ObjectPool at all.  I would need to rewrite ExecuteNodeTreeAsync and ExecuteNodeAsync to implement the ObjectPool in that PR, if the ObjectPool was deemed to be a good choice for the project -- if it could be utilized at all in an effective manner.

Let's look for another use for the ObjectPool.