@sungam3r 2020-03-03
>  For instance, if someone utilized an UnhandledExceptionDelegate to store references to the context of failed resolvers

I would say that this should not be done regardless of my PR. Anyway inside the context there may be references to already destroyed objects. I am sure we should note it in the documentation. 

> Note that this package is not part of .Net Standard 2.0

So what? Netstandard2.0 targeted.

> applications that rely on .Net Core 2.2 (like mine) should be able to safely upgrade that one nuget package without breaking compatibility with the rest of their application.

https://www.fuget.org/packages/Microsoft.Extensions.ObjectPool/3.1.1/lib/netstandard2.0/diff/2.2.0/
No breaking changes, also I can downgrade to 2.2 or even lower. It doesn’t matter to me.

> I'm also confused why you don't seem to mind this dependency, whereas you've called other dependencies "leaky"

Because this dependency is entirely internal, it by definition does not leak anywhere in user code - no interfaces to implement, no classes to use.

> let's compare this to #1535

As an alternative, I agree. Waiting for detailed review.

> the context is only valid during the execution of the resolver, middleware, error handlers, etc

Still, I would mention this in the documentation. The context (in a broad sense) contains some information for the duration of some piece of code (algorithm section). The correctness of the context outside this section is not guaranteed, although sometimes, by a fluke, the context is not destroyed. I can give `HttpContext` as an example. What will happen if you save it and use it somehow further? Probably nothing good.

About DataLoader - sorry, so far this topic is beyond my knowledge.