@BenjaBobs 2020-03-03
> Anyway inside the context there may be references to already destroyed objects. I am sure we should note it in the documentation.
Doesn't hurt, and I think it would add value. I imagine this is one of the cases where if someone tries this it could take them a while to figure out what's happening.