@Shane32 2020-09-13
@sungam3r do you plan to revisit this PR?  I suggest closing it due to (1) complications with #1511, (2) additional requirements on user code, and (3) minimal "benefit".  #1535 does not carry the added requirements or complications, and we have already dismissed that PR as unnecessary.